Fixing cases where password were being displayed
Sometimes the "echo -e "$sudoPass\n" | yes" command
were displaying the password instead of confirming
something. So I changed it to make one thing at a
time.

First I refresh the use of sudo with password,
just using sudo in an echo command. So I can
use the command that really needs it with
not worries about asking the password.