Adding script to check DNS on laptop
Laptop battery is not working, so when the clock is wrong, my DNS
server does not works (because of DNSSEC) and it makes the clock to not
be able to update properly.

So this script changes the DNS to Quad9 and forces update the clock time
before changing it back to my DNS server.